Fix cannot load bug and organise code base
