+++ README.md
@@ -1,4 +1,4 @@
-# Duke project template
+# app.Duke project template
 
 This is a project template for a greenfield Java project. It's named after the Java mascot _Duke_. Given below are instructions on how to use it.
 
@@ -13,7 +13,7 @@ Prerequisites: JDK 11, update Intellij to the most recent version.
    1. If there are any further prompts, accept the defaults.
 3. Configure the project to use **JDK 11** (not other versions) as explained in [here](https://www.jetbrains.com/help/idea/sdk.html#set-up-jdk).<br>
    In the same dialog, set the **Project language level** field to the `SDK default` option.
-4. After that, locate the `src/main/java/Duke.java` file, right-click it, and choose `Run Duke.main()` (if the code editor is showing compile errors, try restarting the IDE).
-   1. Important: In Duke.java, configure isTest to false.
+4. After that, locate the `src/main/java/app.Duke.java` file, right-click it, and choose `Run app.Duke.main()` (if the code editor is showing compile errors, try restarting the IDE).
+   1. Important: In app.Duke.java, configure isTest to false.
 5. To run the IO tests, navigate to the text-ui-test directory and run `./runtest.sh`.
-   1. Important: In Duke.java, configure isTest to true.
+   1. Important: In app.Duke.java, configure isTest to true.
