+++ src/main/java/Launcher.java
@@ -1,3 +1,4 @@
+import app.Main;
 import javafx.application.Application;
 
 /**
@@ -5,6 +6,10 @@ import javafx.application.Application;
  */
 public class Launcher {
     public static void main(String[] args) {
-        Application.launch(Main.class, args);
+        try {
+            Application.launch(Main.class, args);
+        } catch (IndexOutOfBoundsException e) {
+            System.out.println(e.getMessage());
+        }
     }
 }
