--- src/main/java/sample/tasks1.txt
@@ -1,4 +1,0 @@
-T | 1 | read book
-D | 0 | return book | 02/08/2022 1000
-E | 0 | project meeting | Aug 6th 2-4pm
-T | 1 | join sports club
