+++ src/test/java/UiTest.java
@@ -1,3 +1,4 @@
+import ui.Ui;
 import exceptions.NoTasksException;
 import objects.Task;
 import objects.Todo;
